client cleanup
