+++ src/Altinn.App.Core/Internal/AccessManagement/AccessManagementClient.cs
@@ -5,10 +5,8 @@ using Altinn.App.Core.Features;
 using Altinn.App.Core.Internal.AccessManagement.Exceptions;
 using Altinn.App.Core.Internal.AccessManagement.Helpers;
 using Altinn.App.Core.Internal.AccessManagement.Models;
-using Altinn.App.Core.Internal.AccessManagement.Models.Shared;
 using Altinn.App.Core.Internal.App;
 using Altinn.Common.AccessTokenClient.Services;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
@@ -28,14 +26,10 @@ internal sealed class AccessManagementClient(
     Telemetry? telemetry = null
 ) : IAccessManagementClient
 {
-#pragma warning disable CA1822
-    internal void DelegationCheck() { }
-#pragma warning restore CA1822
-
     public async Task<DelegationResponse> DelegateRights(DelegationRequest delegation, CancellationToken ct)
     {
-        // TODO: telemetry
-        var onlyToRemoveWarning = telemetry?.IsInitialized;
+        using var activity = telemetry?.StartAppInstanceDelegationActivity();
+
 
         HttpResponseMessage? httpResponseMessage = null;
         string? httpContent = null;
@@ -45,10 +39,9 @@ internal sealed class AccessManagementClient(
             var application = await appMetadata.GetApplicationMetadata();
 
             var uri = urlHelper.CreateInstanceDelegationUrl(delegation.ResourceId, delegation.InstanceId);
-            AppsInstanceDelegationRequestDto dto = GetDto(delegation);
-            var body = JsonSerializer.Serialize(dto);
+            var body = JsonSerializer.Serialize(DelegationRequest.ConvertToDto(delegation));
             logger.LogInformation(
-                $"Delegating rights to {dto.To.Value} from {dto.From.Value} for {delegation.ResourceId}"
+                $"Delegating rights to {delegation.To?.Value} from {delegation.From?.Value} for {delegation.ResourceId}"
             );
 
             using var httpRequestMessage = new HttpRequestMessage(HttpMethod.Post, uri)
@@ -64,38 +57,13 @@ internal sealed class AccessManagementClient(
             httpResponseMessage = await httpClient.SendAsync(httpRequestMessage, ct);
             httpContent = await httpResponseMessage.Content.ReadAsStringAsync(ct);
             DelegationResponse? response;
-            if (httpResponseMessage.IsSuccessStatusCode)
+            if (!httpResponseMessage.IsSuccessStatusCode)
             {
-                response = JsonSerializer.Deserialize<DelegationResponse>(httpContent);
-                if (response is null)
-                    throw new JsonException("Couldn't deserialize access management response.");
-            }
-            else
-            {
-                try
-                {
-                    var problemDetails = JsonSerializer.Deserialize<ProblemDetails>(httpContent);
-                    if (problemDetails is not null)
-                    {
-                        logger.LogError(
-                            "Got error status code for access management request. Status code: {StatusCode}. Problem details: {ProblemDetails}",
-                            httpResponseMessage.StatusCode,
-                            JsonSerializer.Serialize(problemDetails)
-                        );
-                        throw new AccessManagementRequestException(
-                            "Got error status code for access management request.",
-                            problemDetails,
-                            httpResponseMessage.StatusCode,
-                            httpContent
-                        );
-                    }
-                }
-                catch (JsonException)
-                {
-                    response = null;
-                }
                 throw new HttpRequestException("Got error status code for access management request.");
             }
+            response = JsonSerializer.Deserialize<DelegationResponse>(httpContent);
+            if (response is null)
+                throw new JsonException("Couldn't deserialize access management response.");
             return response;
         }
         catch (Exception e)
@@ -112,47 +80,11 @@ internal sealed class AccessManagementClient(
                     );
             logger.LogError(ex, "Error when processing access management request.");
 
-            // TODO: metrics
-
             throw ex;
         }
         finally
         {
             httpResponseMessage?.Dispose();
         }
-    }
-
-    private static AppsInstanceDelegationRequestDto GetDto(DelegationRequest delegation)
-    {
-        return new AppsInstanceDelegationRequestDto
-        {
-            From = new DelegationParty
-            {
-                Type = delegation.From is not null
-                    ? delegation.From.Type
-                    : throw new AccessManagementArgumentException("From is required"),
-                Value = delegation.From.Value,
-            },
-            To = new DelegationParty
-            {
-                Type = delegation.To is not null
-                    ? delegation.To.Type
-                    : throw new AccessManagementArgumentException("To is required"),
-                Value = delegation.To.Value,
-            },
-            Rights = delegation
-                .Rights.Select(r => new RightDto
-                {
-                    Resource = r.Resource.Select(rr => new Resource { Type = rr.Type, Value = rr.Value }).ToList(),
-                    Action = new AltinnAction
-                    {
-                        Type = r.Action is not null
-                            ? r.Action.Type
-                            : throw new AccessManagementArgumentException("Action is required"),
-                        Value = r.Action.Value,
-                    },
-                })
-                .ToList(),
-        };
     }
 }
